modify testcode
